Opened a window, began parsing
1. main.c currently is able to open a window with a square. That's about
it. The rgb() function takes three integers in rgb format and will
convert them to a single integer appropriate for the mlx_put_pixel()
function. ft_hex is new to libft.a and can convert hex strings into
integers, if hexadecimal input is desired. (MAKE ANOTHER FUNCTION
ANALOGOUS TO rgb() THAT ALLOWS INPUT OF A SINGLE HEX STRING).

2. main2.c is beginning the project, with parsing efforts coming first.
Structures are defined in the cub3d.h header file and should be used
extensively through the project. The first target is to parse all input
correctly into a structure or structures, then think of what comes next.