Improved error messaging and beginning of parsing
1) Parsing functions have been written, and should have sufficient error
checking. However, both parsing and error detection are not yet tested.

2) An ft_printf_fd function would be really useful, but altering my
existing ft_printf_fd will be really difficult, due to the code being
very unclean. It might be sensible to create an ft_printf_fd based on
the miniprintf projects that I completed, although it wouldn't have as
much functionality as my complete ft_printf, it is good enough for many
purposes.