Improved error messaging and map parsing enabled
1) The cub3d libft has been modified, with a new version of ft_printf
added, ft_printf_fd. This miniprintf is not as effective as the full
ft_printf, as modifications to the real ft_printf are made difficult by
the messiness of the code. ft_printf_fd however can handle 'd', 'x' and
's', width and precision, and unlike ft_printf, ft_printf can be used to
print to STDERR or to any open file by using it's file descriptor.

2) ft_printf_fd has been used effectively in ft_error to output not just
an error message, but the line that is incorrectly formatted in the .cub
file.

3) Parsing of .cub files is complete, and all map information, texture
information, colours, and player location + direction are stored in
structures (see cub3d.h). Error checking is constant throughout the
parsing process, and exit() calls are present utilising ft_error.
However, a function to free a char ** is required, to free the only
heap-allocated memory in the stack, which is the map.

Next goals are to write this freeing function to be utilised by exit()
calls, and then to begin the raycasting process.