Raycasting improved and image work beginning
Images are going to be essential for the raycasting to work at a
reasonable speed, currently attempting to draw to an image by using its
memory address, but so far I'm not having great success in either
editing the pixels of the image or in putting the image to a window.

Raycasting function needs fixing for norminette as the function is
around 80 lines, but the basic logic is working well now. The plane
vector has been calculated properly now using sin and cos.