Basic functionality achieved
Images are working correctly, and are generated with good speed.
Movement is possible in all 4 directions, and rotation left or right.
Basic collision detection is present as well, and it is not possible to
escape from the map.

Next goals should be to get textures active, go for sprites too, and
then clean up the code, make norm-compliant etc.

Following norm compliance, the plan is to carefully read and re-read the
subject page for cub3d, and check if I'm missing anything - see what the
bonuses are and if they are achievable.