Crashes fixed, movement improved, norm improved.
- In last commit, the fix to un-mirror textures resulted in texX gaining
a value of 128, outside of the range of 0 to 127. This resulted in
common segmentation faults. Fixed easily by adding a '- 1' in
raycasting.

- Movement improved in move_forward, soon to be applied to other
functions as well. Previously, movement was simply testing whether the
next space was free, now a short dda check is performed to check if
there is anything in the way first to prevent walking through corners.
Movement needs to be improved in two further ways in future: allowing
'sliding' along walls, using only the component of movement that is
possible. Secondly, movement should begin and loop immediately following
a key press, and end after release of that key.

- ray(), which was 160 lines, has been chopped into pieces to fit the
norm. The large list of variables all now fits within a structure t_ray.
t_ray includes another structure, t_dda, that has been very useful in
the improved movement as well.