New movement
After spending a lot of time on DDA-based movement, I saw Duncan's
implementation of movement, and after checking the original Wolfenstein
it turns out that is the implementation the original uses as well. Not
only is it better, it's also much simpler, making me feel a bit dumb
about spending so much time on DDA-based movement.

This should be the last commit with this type of movement, but we can
come back here if any issues arise in future.