Overhaul completed
Functionality is now as good as ever, textures work correctly for
ceiling and floor, as well as all the walls. Functions have been
rearranged into separate files with relevant names - and everything is
now norm-compliant. There may be a couple of old things in the cub3d.h
header file that could be changed, as structures have undergone a
complete makeover.

Overall, the code is much cleaner, more readable, and more easily
extended than before. The next step is to add sprites and then decide
however many of the bonus challenges I would like to take on.