lldb saves my life
Very stupid mistake that took hours to find: in img_put_pixel, the 'if'
statement 'if (x < 0 && y < 0)' should have been 'if (x < 0 || y < 0)'.
Shouldn't really be receiving negative y values anyway, but clearly it
was in very rare cases.

Sprites now fully working, if a little inefficient.
Priority now should be exporting image to file, and then the basic
requirements of the project are completed.