Djevayo mock eval finds a lot of errors
See notebook for errors that have been found - unfortunately there is
quite a few, and most are due to the parsing of the file.

Memory leaks, segmentation faults if one of the parameters is missing,
and ignoring of digits that occur after the normal digits in get_color,
lots to do.