Errors mostly fixed, only bitmap export remaining
Every error that was found by Djevayo has been fixed, parsing of the
file is greatly improved now, and no longer susceptible to incorrect
input resulting in segmentation faults. This doesn't mean that there
aren't other errors that have yet to be found however.

Bitmap export now needs to be written, and in particular I need to
ensure that padding is completed in the case of strange resolutions such
as 550 x 550.