Couple issues remaining
bmp file can now be created, and written to, but doesn't fill in any of
the pixel information yet.

I need to change the way that my image works, because currently it uses
increasing amounts of memory (even though the image creation and freeing
is working perfectly), might be a bit annoying to fix, especially
considering the issue that others have had with sprites flickering.