Map validation completed
Flood fill algorithm is now working correctly, if there are any other
issues, I haven't been able to find them.
I changed the map validation to fit in with Jonas' interpretation of the
subject, simply checking to see that all external walls of the map are
also '1's to ensure "the map is closed/surrounded by walls".

This commit has a few functions that are helpful to make the flood fill
algorithm very clear, however a great deal of tidying is now to be done
to make the project norm-compliant.

Bonuses can wait for a while, I'm going to submit with the bonuses I
have and I can add extra functionality later on.