Minor change to allow 0's outside of playermap
Jonas was convinced that the 'map' definition in the subject meant the
entire 2D array, I thought it was just the player's zone and according
to the creator of cub3d my initial thought was correct.

An if statement that renders certain types of maps invalid was left in
by accident while reverting back to the 'entire map' validity check,
however, it's changed back in this commit.