Error from first eval fixed
An oversight allowed sprites to have a value of 0 when very small
resolutions was used. Unfortunately, this would result in a
floating-point exception crash, as the program would divide by this
value.
Fixed by checking to see if sprite width is less than 0, and if so,
setting the value to equal 1 (found in sprite_cast.c).