Edits after eval
My map wasn't dealing well with files that didn't end in a new line (as
they always do when you use a proper text editor like vim), so VSC tests
messed things up a bit. Not sure if this code will accept new lines
after the map, might need to fix it so it does (but doesn't accept
anything but new lines).