Threading actually fully implemented
Threading was actually only active on the sprites before, now it is
active on walls and floors also.

Speed is really great, rendering is done very quick unless in the case
of many many sprites - should perform final checks tomorrow for crashes
on weird resolutions, odd parsing errors, and then should be ready to
submit.