Couple minor tweaks
An incorrect error message was returned if the player was on the very
edge of an incorrectly contained map, as the while loops in find player
were actually missing the final index of the strlen or arrlen.
Would never result in a crash, but the incorrect error message was
given.
Fixed now.

Threading also protected in bonus files.