Many modifications made to run on home computer
- Swift, and particularly the 'metal' library are not present on my home
laptop, and impossible to get, so I found the miniLibX that uses solely
C from Thijs' git repo, and cloned it here.

- A lot of modifications were required to the Makefile to correctly make
with this new library - for some reason gcc in particular functions very
differently on my home computer, and requires that the -L and -l come
after the name of the file to be produced, and the object files, when
linking. The 'make noswift' command now will create the executable in a
format suitable for my home computer.

- Using valgrind, I discovered an error in my code, where my
initialisation of 'info' was actually doing nothing at all, as I was
creating a new info structure in parse_cub(), and reassigning the
initialised info structure. As a result, there were multiple instances
of uninitialised variables being used in conditional statements, which I
think is fixed now by simply using the original info structure that has
been initialized effectively.

- An '#ifdef' and '#ifndef' are used in my cub3d.h now to ensure that
the key codes are correct for whichever system I am on, Linux or Mac
OSX.

- Multiplied the threads in sprites by two again to improve performance
further.

- TO-DO
	- My makefile is okay, but it could be improved further for
	greater readability, and particularly to detect the relevant OS
	and compile accordingly. I may look to Thijs' Makefile for
	reference, as it is very instructional.
	- Norminette check again.
	- Improve the sorting algorithm for sprites, as bubble sort is
	one of the very slowest methods possible and is significantly
	affecting performance on maps with a large sprite count.
	- Double check that the threading for sprites will not be used
	at low screen resolutions.