Home changes done
A couple extra textures added and maps changed. I think that there are
issues in the bitmap creation currently, that I made while trying to
compile on my home laptop.

The Makefile definitely needs some more work currently to make
compilation easier and not require the instruction 'noswift' every time
I compile on my home computer.

There may be a couple of norminette errors also.

These are just some small things that should be fixed when I come back
to work on this project in the future - adding bonuses etc.